refactor: Extract VisualCanvas from WorkZone, eliminate prop-drilling
Extract ~3000 lines of canvas rendering logic from WorkZone.tsx (4126→1047 lines)
into reusable components and hooks that read directly from the Zustand store:

- canvasHelpers.ts: Pure functions for dimensions, themes, colors, link logic
- useCanvasDragDrop.ts: All drag-drop handlers reading/writing store
- useCanvasResize.ts: Block resize handlers with collaboration sync
- VisualCanvas.tsx: Section drop zone wrapper component
- PlaySectionContent.tsx: Moved to canvas/, props reduced from 25+ to 2
- BlockSectionContent.tsx: Moved to canvas/, props reduced from 25+ to 3
- playbookEditorStore.ts: Added toggle actions (block collapse, sections)

Includes 35 unit tests for canvasHelpers. All 181 tests pass.
